Apply rust review batch 15: partial sweep writes, doc homes, env trim
- the version check writes only the columns it sets, so a row read at the
  top of a sweep minutes long cannot overwrite a build's status change
  that landed meanwhile
- integer env settings trim whitespace before parsing (+ test)
- best-effort cache tasks on the worker log their panics instead of
  dropping them (+ test)
- doc comments moved back onto the items they describe; the rest of the
  settings routes and `list_builds` get one for the OpenAPI document
- the client decodes the restore response like every other JSON answer;
  `PackageCandidate` reads straight off the row; OAuth uses `bearer_auth`
- the dashboard adds package counts wide, like the success rate

+++ frontend-rs/src/screens/dashboard.rs
@@ -98,7 +98,9 @@ fn format_rate(rate: Option<f64>) -> String {
 /// The headline numbers.
 #[component]
 fn StatTiles(stats: ListStats) -> Element {
-    let all_packages = stats.requested_packages + stats.dependency_packages;
+    // Added wide, like `success_rate` below: the behaviour on overflow should
+    // not differ between two tiles of the same page.
+    let all_packages = u64::from(stats.requested_packages) + u64::from(stats.dependency_packages);
 
     rsx! {
         div { class: "grid gap-4 grid-cols-2 lg:grid-cols-3 xl:grid-cols-5",
